+++ admin-dev/themes/new-theme/js/components/form/form-object-mapper.ts
@@ -209,6 +209,9 @@ export default class FormObjectMapper {
     // will be filled respecting the order in the DOM
     const inputs: Array<HTMLElement> = [];
     const domForm = this.$form.get(0);
+
+    if (!domForm) return;
+
     inputNames.forEach((inputName: string) => {
       const inputsByName = domForm.querySelectorAll(`[name="${inputName}"]`);
