+++ admin-dev/themes/new-theme/js/types/vue/shim.d.ts
@@ -1,0 +1,5 @@
+declare module '*.vue' {
+  import {defineComponent} from 'vue';
+
+  export default defineComponent;
+}
