+++ admin-dev/themes/new-theme/package.json
@@ -93,7 +93,7 @@
     "node-sass": "^6.0.1",
     "path": "^0.12.7",
     "postcss-loader": "^5.3.0",
-    "postcss-preset-env": "^6.7.0",
+    "postcss-preset-env": "^7.7.1",
     "sass-loader": "^12.1.0",
     "style-loader": "^1.1.3",
     "stylelint": "^13.13.1",
@@ -107,7 +107,7 @@
     "vue-eslint-parser": "^9.0.3",
     "vue-loader": "^17.0.0",
     "vue-style-loader": "^4.1.3",
-    "webpack": "^5.53.0",
+    "webpack": "^5.73.0",
     "webpack-cli": "^4.7.0",
     "webpack-dev-server": "^4.3.0",
     "webpack-font-preload-plugin": "^1.4.1",
